Fix concurrent-modification bug in connectionClosed.
Since the streamManager is generally shared among all connections of a given
store, and connections be used in different threads, the list of open streams
may be modified in one thread while it's being iterated over in another
(creating a copy of the Set involves iterating over that Set). So the copy
must be done while synchronized on the Set.