Fix (most) deadlocks by making sure indexes are used for all updates/deletes. This is because during modifications Derby acquires locks on potentially untouched rows (even in READ_UNCOMMITTED mode) if it uses a table-scan to find the rows to be modified. So by forcing it to use indexes all the time we can avoid (most) of these issues.
Unfortunately there is a still a rare case where it could end up deadlocking,
in which case Derby's lock timeout will kick us out with an exception, meaning
the transaction will have to be aborted and retried. To avoid this completely,
an optional flag may be specified to force all writers to be serialized, i.e.
allowing only a single write-transaction at a time.

Furthermore, Derby 10.4 (and earlier) has a bug that cause Derby to acquire
locks when it shouldn't; I've tested with 10.5.1.0 (beta) and have not
encountered any locking issues while using it, however. Therefore by default
the single-writer flag is set if using Derby version prior to 10.5, and not set
if using 10.5 or later.