Reverse the call-back to Transaction from the akubra-rmi-server side. A listener to accept calls and enque for client side to service is the general concept. This also allows all calls to the Transaction object on the akubra-client-side to be made on the same thread context as the one making the BlobStore.openConnection().
All calls to the Transaction object on the akubra-server-side after
the openConnection() call returns is returned with an
IllegalStateException. This is allowed as per JTA specs since
the thread context is most definitely changed.

Note also that calls to XAResource objects as part of the
txn.enlistResource() will be on a different thread. XAResource
objects are expected to handle multi threaded operations as
per JTA specs.

With this all limitations mentioned in r2617 is addressed.