Created a basic TCK, a test suite aiming to provide a fairly complete set of tests for the core Akubra functionality that each store implementation must provide in order to ensure the Akubra API's are properly, consistently, and fully implemented.
All current stores except the rmi store have had their tests modified or
augmented with the TCK test suit; doing this for the rmi-store is left as an
exercise for the reader. In many cases the original tests were left almost
untouched; more removing of duplicate tests and/or merging with the TCK tests
is probably desired.

The current test suit is basically one large class. I've considered splitting
this into, say, a store, a connection, and a blob test suite class. While this
may seem more modular, the disadvantage is that certain overriding of methods
that most tests need to do would have to be duplicated for each class.