AKUBRA-7: FSBlobIdIterator escapes file paths
Path components are no longer escaped during iteration as this was
causing problems using akubra-fs to work with files that contained
characters like '+' and ':'. Having the iterator escape path components
was a misguided feature aimed at allowing akubra-fs to work with arbitrary
file paths. In fact, it is unreasonable for akubra-fs to work with non-URI
compatible paths because of the ambiguities that arise in practice. So the
iterator now ignores such paths with a warning when they are encountered.

Related email thread:
http://groups.google.com/group/akubra-dev/browse_thread/thread/cebc61ea517d7c90

https://jira.duraspace.org/browse/AKUBRA-7