AKUBRA-8: Update libs for 0.4.0
This commit updates all dependencies to the latest released
version on central as of March 2nd, 2012, except where explicitly
noted in comments in the root pom. All compile and runtime dependencies
were changed without requiring code updates. However, for running tests,
the easymock and bitronix updates did require minor test code changes,
and the logback upgrade required a minor change in the way the logback.xml
files are constructed.

In addition, this updates all version declarations to be explicitly
made in the root pom. This applies to plugins as well.

https://jira.duraspace.org/browse/AKUBRA-8